// apps/web/lib/api/types/PublicCollection.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { PublicItemView } from "./PublicItemView"
import type { SetCoverage } from "./SetCoverage"
import type { TypeCount } from "./TypeCount"

/**
 * PublicCollection
 */
export type PublicCollection = {
  /**
   * @type integer
   */
  total_cards: number
  /**
   * @type integer
   */
  total_groups: number
  /**
   * @type array
   */
  types: TypeCount[]
  /**
   * @type array
   */
  sets: SetCoverage[]
  /**
   * @type array
   */
  items: PublicItemView[]
}